refactor: move mock context to own file
